fix: remove inline styles for verifyButton
